# Use System.Buffers.ArrayPool