--- a/libZipSharp.csproj
+++ b/libZipSharp.csproj
@@ -55,5 +55,6 @@
     </ItemGroup>
     <ItemGroup>
       <PackageReference Include="Mono.Posix.NETStandard" Version="1.0.0" PrivateAssets="none" />
+      <PackageReference Include="System.Buffers" Version="4.5.0" />
     </ItemGroup>
 </Project>