@dellis1972 2020-02-24
This looks good. I am beginning to wonder if those Utf8String marshalling methods might be more performant using `Span<byte>` to process the native memory directly, rather than copying it over.
That said the `Span.ToArray` seems to do a managed creation of an array anyway. So it won't buy us anything.

Or perhaps even better?

```
public unsafe static string Utf8StringPtrToString (IntPtr utf8Str)
        {
            if (utf8Str == IntPtr.Zero)
				return null;

			byte* ptr = (byte*)utf8Str.ToPointer ();
			byte* p = ptr;
			while (*p != 0)
				p++;
			var len = (int)(p - ptr);
			if (len == 0)
				return String.Empty;

            return Encoding.UTF8.GetString ((byte*)utf8Str.ToPointer (), len);
        }
```

Just use the `UTF8.GetString (byte* ptr, int len)` override instead and do away with the buffers completely?