@dellis1972 2020-02-24
@jonathanpeppers  I've been playing with System.Buffers on another app, and I'm getting a bunch of dependency issues with netcore 3.1. It keeps trying to load System.Buffers 4.0.2.0 instead of the 4.5.0 one I've specified. Also I can't seem to get it to include the dll in the local build. Are we sure this is not going to break anything?