@jonathanpeppers 2020-02-24
When I tested with `xamarin-android`, I did this:
```
$ msbuild libZipSharp.csproj /p:Configuration=Release
$ cp bin/Release/netstandard2.0/libZipSharp.dll ../xamarin-android/bin/Release/lib/xamarin.android/xbuild/Xamarin/Android/
```
I could see my changes working, because I got different stack traces in the profiler.

Is there an easy way to try out the nupkg on a `xamarin-android` PR build? I can do it locally, but it won't be exactly the same as doing a full PR build that tests everything.